feat :: 백그라운드 => 포그라운드 시 갱신 로직 위치 이동
main_page에서 bottom_navi로 이동시켜 모든 화면에서 갱신이 가능하도록 변경함
